+++ hutool-extra/src/main/java/cn/hutool/extra/cglib/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Cglib库方法封装
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.extra.cglib;